+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [66.4.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@66.4.0...@codecademy/gamut@66.4.1) (2025-09-15)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+## [66.4.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@66.3.1...@codecademy/gamut@66.4.0) (2025-09-12)
+
+### Features
+
+- add abbreviations to SelectDropdown, fixed grouped select + formatting issues ([1f01312](https://github.com/Codecademy/gamut/commit/1f013125d104d24d7b9f9b1fca275a49812ebda0))
+
 ### [66.3.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@66.3.0...@codecademy/gamut@66.3.1) (2025-09-09)
 
 **Note:** Version bump only for package @codecademy/gamut